+++ clients/ops/admin-ui/src/constants.ts
@@ -26,6 +26,14 @@ export const USER_PRIVILEGES: UserPrivileges[] = [
     scope: "connection:create_or_update",
   },
   {
+    privilege: "Instantiate connections to SaaS datastores",
+    scope: "connection:instantiate",
+  },
+  {
+    privilege: "Read connection types",
+    scope: "connection_type:read",
+  },
+  {
     privilege: "Delete datastore connections",
     scope: "connection:delete",
   },